Check in to fix change #26500 for Open Watcom tree
[git-p4: depot-paths = "//depot/openwatcom/": change = 26501]

+++ bld/setupgui/h/dynarray.h
@@ -31,13 +31,13 @@
 
 
 typedef struct array_info {
-    int         num;
-    int         alloc;
-    int         esize;
+    size_t      num;
+    size_t      alloc;
+    size_t      esize;
     void        **array;
-    int         increment;
+    size_t      increment;
 } array_info;
 
-extern void InitArray( void **array, int esize, array_info *info );
+extern void InitArray( void **array, size_t esize, array_info *info );
 extern bool BumpArray( array_info *info );
 extern bool BumpDownArray( array_info *info );
